# ai-readiness-frontend/lib/openapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Readiness Assessment API",
    "description": "Comprehensive API for AI readiness assessment platform including survey management, analytics, LLM analysis, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "AI Readiness Support",
      "email": "support@ai-readiness.com",
      "url": "https://ai-readiness.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://ai-readiness-frontend.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/analytics/dashboard": {
      "get": {
        "summary": "Get dashboard analytics",
        "description": "Retrieve comprehensive analytics data for dashboard display",
        "operationId": "getDashboardAnalytics",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "description": "Organization ID to filter results",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date for data range",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date for data range",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "metrics",
            "in": "query",
            "description": "Comma-separated list of metrics to include",
            "schema": {
              "type": "string"
            },
            "example": "overview,jtbd_trends,voice_quality"
          },
          {
            "name": "period",
            "in": "query",
            "description": "Aggregation period",
            "schema": {
              "type": "string",
              "enum": ["daily", "weekly", "monthly"],
              "default": "weekly"
            }
          },
          {
            "name": "export",
            "in": "query",
            "description": "Export data instead of displaying",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Export format",
            "schema": {
              "type": "string",
              "enum": ["json", "csv"],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAnalytics"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                },
                "example": "overview,{\"totalResponses\":1247,\"completionRate\":87.5}"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Advanced analytics query",
        "description": "Submit complex analytics queries with custom filters and parameters",
        "operationId": "advancedAnalyticsQuery",
        "tags": ["Analytics"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvancedAnalyticsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Advanced analytics results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAnalytics"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/llm/analyze": {
      "post": {
        "summary": "Analyze response with LLM",
        "description": "Perform Jobs-to-be-Done (JTBD) analysis on survey responses using Large Language Models",
        "operationId": "analyzeLLMResponse",
        "tags": ["LLM Analysis"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMAnalysisRequest"
              },
              "example": {
                "responseId": "uuid-response-id",
                "responseText": "We struggle with data quality and need better integration tools",
                "questionText": "What are your main challenges with current systems?",
                "expectedForce": "pain_of_old",
                "questionContext": "AI readiness assessment",
                "organizationId": "uuid-org-id",
                "surveyId": "uuid-survey-id"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "LLM analysis completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - LLM access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "LLM service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "LLM service health check",
        "description": "Check the status and health of LLM analysis services",
        "operationId": "llmHealthCheck",
        "tags": ["LLM Analysis"],
        "responses": {
          "200": {
            "description": "LLM service status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMHealthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Supabase authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "DashboardAnalytics": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "overview": {
                "$ref": "#/components/schemas/OverviewMetrics"
              },
              "jtbd_trends": {
                "$ref": "#/components/schemas/JTBDMetrics"
              },
              "voice_quality": {
                "$ref": "#/components/schemas/VoiceMetrics"
              },
              "user_engagement": {
                "$ref": "#/components/schemas/EngagementMetrics"
              },
              "real_time": {
                "$ref": "#/components/schemas/RealTimeMetrics"
              }
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "organizationId": {
                "type": "string",
                "nullable": true
              },
              "dateRange": {
                "type": "object",
                "properties": {
                  "start": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "end": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                }
              },
              "period": {
                "type": "string",
                "enum": ["daily", "weekly", "monthly"]
              },
              "metrics": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "cacheTtl": {
                "type": "number",
                "description": "Cache TTL in seconds"
              }
            }
          }
        }
      },
...
      "RealTimeMetrics": {
        "type": "object",
        "properties": {
          "currentUsers": {
            "type": "number"
          },
          "activeSurveys": {
            "type": "number"
          },
          "systemHealth": {
            "type": "number",
            "format": "float",
            "description": "System health percentage"
          },
          "responseRate": {
            "type": "number",
            "format": "float",
            "description": "Real-time response rate"
          }
        }
      },
      "AdvancedAnalyticsRequest": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "dateRange": {
            "type": "object",
            "properties": {
              "start": {
                "type": "string",
                "format": "date-time"
              },
              "end": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "metrics": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["overview", "jtbd_trends", "voice_quality", "user_engagement", "real_time"]
            },
            "default": ["overview"]
          },
          "period": {
            "type": "string",
            "enum": ["daily", "weekly", "monthly"],
            "default": "weekly"
          },
          "export": {
            "type": "boolean",
            "default": false
          },
          "format": {
            "type": "string",
            "enum": ["json", "csv"],
            "default": "json"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["error"]
      },
...
    }
  },
...
}